+++ event-sourcing/subscription-manager/src/test/java/uk/gov/justice/services/event/sourcing/subscription/manager/SubscriptionEventProcessorTest.java
@@ -40,7 +40,6 @@ import uk.gov.justice.services.metrics.micrometer.counters.MicrometerMetricsCoun
 
 import java.util.UUID;
 
-import javax.transaction.UserTransaction;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -75,9 +74,6 @@ public class SubscriptionEventProcessorTest {
 
     @Mock
     private MicrometerMetricsCounters micrometerMetricsCounters;
-
-    @Mock
-    private UserTransaction userTransaction;
 
     @Mock
     private TransactionHandler transactionHandler;
@@ -134,7 +130,7 @@ public class SubscriptionEventProcessorTest {
                 micrometerMetricsCounters);
 
         inOrder.verify(micrometerMetricsCounters).incrementEventsProcessedCount(source, component);
-        inOrder.verify(transactionHandler).begin(userTransaction);
+        inOrder.verify(transactionHandler).begin();
         inOrder.verify(newStreamStatusRepository).lockStreamAndGetStreamUpdateContextWithError(
                 streamId,
                 source,
@@ -145,10 +141,10 @@ public class SubscriptionEventProcessorTest {
         inOrder.verify(newEventBufferRepository).remove(streamId, source, component, eventPositionInStream);
         inOrder.verify(newStreamStatusRepository).setUpToDate(true, streamId, source, component);
         inOrder.verify(micrometerMetricsCounters).incrementEventsSucceededCount(source, component);
-        inOrder.verify(transactionHandler).commit(userTransaction);
+        inOrder.verify(transactionHandler).commit();
 
         verify(streamErrorRepository, never()).markStreamAsFixed(any(UUID.class), eq(streamId), eq(source), eq(component));
-        verify(transactionHandler, never()).rollback(userTransaction);
+        verify(transactionHandler, never()).rollback();
         verify(micrometerMetricsCounters, never()).incrementEventsFailedCount(source, component);
     }
 
@@ -198,7 +194,7 @@ public class SubscriptionEventProcessorTest {
                 micrometerMetricsCounters);
 
         inOrder.verify(micrometerMetricsCounters).incrementEventsProcessedCount(source, component);
-        inOrder.verify(transactionHandler).begin(userTransaction);
+        inOrder.verify(transactionHandler).begin();
         inOrder.verify(newStreamStatusRepository).lockStreamAndGetStreamUpdateContextWithError(
                 streamId,
                 source,
@@ -209,10 +205,10 @@ public class SubscriptionEventProcessorTest {
         inOrder.verify(newEventBufferRepository).remove(streamId, source, component, eventPositionInStream);
         inOrder.verify(newStreamStatusRepository).setUpToDate(true, streamId, source, component);
         inOrder.verify(micrometerMetricsCounters).incrementEventsSucceededCount(source, component);
-        inOrder.verify(transactionHandler).commit(userTransaction);
+        inOrder.verify(transactionHandler).commit();
 
         verify(streamErrorRepository, never()).markStreamAsFixed(any(UUID.class), eq(streamId), eq(source), eq(component));
-        verify(transactionHandler, never()).rollback(userTransaction);
+        verify(transactionHandler, never()).rollback();
         verify(micrometerMetricsCounters, never()).incrementEventsFailedCount(source, component);
     }
 
@@ -263,7 +259,7 @@ public class SubscriptionEventProcessorTest {
                 micrometerMetricsCounters);
 
         inOrder.verify(micrometerMetricsCounters).incrementEventsProcessedCount(source, component);
-        inOrder.verify(transactionHandler).begin(userTransaction);
+        inOrder.verify(transactionHandler).begin();
         inOrder.verify(newStreamStatusRepository).lockStreamAndGetStreamUpdateContextWithError(
                 streamId,
                 source,
@@ -275,7 +271,7 @@ public class SubscriptionEventProcessorTest {
         inOrder.verify(streamErrorRepository).markStreamAsFixed(previousStreamErrorId, streamId, source, component);
         inOrder.verify(newStreamStatusRepository).setUpToDate(true, streamId, source, component);
         inOrder.verify(micrometerMetricsCounters).incrementEventsSucceededCount(source, component);
-        inOrder.verify(transactionHandler).commit(userTransaction);
+        inOrder.verify(transactionHandler).commit();
 
         verify(micrometerMetricsCounters, never()).incrementEventsFailedCount(source, component);
     }
@@ -327,7 +323,7 @@ public class SubscriptionEventProcessorTest {
                 micrometerMetricsCounters);
 
         micrometerMetricsCounters.incrementEventsProcessedCount(source, component);
-        inOrder.verify(transactionHandler).begin(userTransaction);
+        inOrder.verify(transactionHandler).begin();
         inOrder.verify(newStreamStatusRepository).lockStreamAndGetStreamUpdateContextWithError(
                 streamId,
                 source,
@@ -337,10 +333,10 @@ public class SubscriptionEventProcessorTest {
         inOrder.verify(newStreamStatusRepository).updateCurrentPosition(streamId, source, component, eventPositionInStream);
         inOrder.verify(newEventBufferRepository).remove(streamId, source, component, eventPositionInStream);
         inOrder.verify(micrometerMetricsCounters).incrementEventsSucceededCount(source, component);
-        inOrder.verify(transactionHandler).commit(userTransaction);
+        inOrder.verify(transactionHandler).commit();
 
         verify(streamErrorRepository, never()).markStreamAsFixed(any(UUID.class), eq(streamId), eq(source), eq(component));
-        verify(transactionHandler, never()).rollback(userTransaction);
+        verify(transactionHandler, never()).rollback();
         verify(newStreamStatusRepository, never()).setUpToDate(true, streamId, source, component);
         verify(micrometerMetricsCounters, never()).incrementEventsFailedCount(source, component);
     }
@@ -390,7 +386,7 @@ public class SubscriptionEventProcessorTest {
                 micrometerMetricsCounters);
 
         inOrder.verify(micrometerMetricsCounters).incrementEventsProcessedCount(source, component);
-        inOrder.verify(transactionHandler).begin(userTransaction);
+        inOrder.verify(transactionHandler).begin();
         inOrder.verify(newStreamStatusRepository).lockStreamAndGetStreamUpdateContextWithError(
                 streamId,
                 source,
@@ -400,10 +396,10 @@ public class SubscriptionEventProcessorTest {
         inOrder.verify(newStreamStatusRepository).updateCurrentPosition(streamId, source, component, eventPositionInStream);
         inOrder.verify(newEventBufferRepository).remove(streamId, source, component, eventPositionInStream);
         inOrder.verify(micrometerMetricsCounters).incrementEventsSucceededCount(source, component);
-        inOrder.verify(transactionHandler).commit(userTransaction);
+        inOrder.verify(transactionHandler).commit();
 
         verify(streamErrorRepository, never()).markStreamAsFixed(any(UUID.class), eq(streamId), eq(source), eq(component));
-        verify(transactionHandler, never()).rollback(userTransaction);
+        verify(transactionHandler, never()).rollback();
         verify(micrometerMetricsCounters, never()).incrementEventsFailedCount(source, component);
     }
 
@@ -449,20 +445,20 @@ public class SubscriptionEventProcessorTest {
                 micrometerMetricsCounters);
 
         micrometerMetricsCounters.incrementEventsProcessedCount(source, component);
-        inOrder.verify(transactionHandler).begin(userTransaction);
+        inOrder.verify(transactionHandler).begin();
         inOrder.verify(newStreamStatusRepository).lockStreamAndGetStreamUpdateContextWithError(
                 streamId,
                 source,
                 component,
                 eventPositionInStream);
-        inOrder.verify(transactionHandler).commit(userTransaction);
+        inOrder.verify(transactionHandler).commit();
 
         verify(interceptorChainProcessor, never()).process(interceptorContext);
         verify(newStreamStatusRepository, never()).updateCurrentPosition(streamId, source, component, eventPositionInStream);
         verify(newEventBufferRepository, never()).remove(streamId, source, component, eventPositionInStream);
         verify(streamErrorRepository, never()).markStreamAsFixed(any(UUID.class), eq(streamId), eq(source), eq(component));
         verify(newStreamStatusRepository, never()).setUpToDate(true, streamId, source, component);
-        verify(transactionHandler, never()).rollback(userTransaction);
+        verify(transactionHandler, never()).rollback();
         verify(micrometerMetricsCounters, never()).incrementEventsSucceededCount(source, component);
     }
 
@@ -508,21 +504,21 @@ public class SubscriptionEventProcessorTest {
                 micrometerMetricsCounters);
 
         inOrder.verify(micrometerMetricsCounters).incrementEventsProcessedCount(source, component);
-        inOrder.verify(transactionHandler).begin(userTransaction);
+        inOrder.verify(transactionHandler).begin();
         inOrder.verify(newStreamStatusRepository).lockStreamAndGetStreamUpdateContextWithError(
                 streamId,
                 source,
                 component,
                 eventPositionInStream);
         inOrder.verify(micrometerMetricsCounters).incrementEventsIgnoredCount(source, component);
-        inOrder.verify(transactionHandler).commit(userTransaction);
+        inOrder.verify(transactionHandler).commit();
 
         verify(interceptorChainProcessor, never()).process(interceptorContext);
         verify(newStreamStatusRepository, never()).updateCurrentPosition(streamId, source, component, eventPositionInStream);
         verify(newEventBufferRepository, never()).remove(streamId, source, component, eventPositionInStream);
         verify(streamErrorRepository, never()).markStreamAsFixed(any(UUID.class), eq(streamId), eq(source), eq(component));
         verify(newStreamStatusRepository, never()).setUpToDate(true, streamId, source, component);
-        verify(transactionHandler, never()).rollback(userTransaction);
+        verify(transactionHandler, never()).rollback();
         verify(micrometerMetricsCounters, never()).incrementEventsFailedCount(source, component);
         verify(micrometerMetricsCounters, never()).incrementEventsSucceededCount(source, component);
     }
@@ -579,19 +575,19 @@ public class SubscriptionEventProcessorTest {
                 micrometerMetricsCounters);
 
         inOrder.verify(micrometerMetricsCounters).incrementEventsProcessedCount(source, component);
-        inOrder.verify(transactionHandler).begin(userTransaction);
+        inOrder.verify(transactionHandler).begin();
         inOrder.verify(newStreamStatusRepository).lockStreamAndGetStreamUpdateContextWithError(
                 streamId,
                 source,
                 component,
                 eventPositionInStream);
         inOrder.verify(interceptorChainProcessor).process(interceptorContext);
-        inOrder.verify(transactionHandler).rollback(userTransaction);
+        inOrder.verify(transactionHandler).rollback();
         inOrder.verify(streamErrorStatusHandler).onStreamProcessingFailure(eventJsonEnvelope, nullPointerException, source, component, streamUpdateContext);
 
         verify(newStreamStatusRepository, never()).updateCurrentPosition(streamId, source, component, eventPositionInStream);
         verify(newEventBufferRepository, never()).remove(streamId, source, component, eventPositionInStream);
-        verify(transactionHandler, never()).commit(userTransaction);
+        verify(transactionHandler, never()).commit();
         verify(micrometerMetricsCounters, never()).incrementEventsSucceededCount(source, component);
         verify(micrometerMetricsCounters, never()).incrementEventsIgnoredCount(source, component);
     }
@@ -634,7 +630,7 @@ public class SubscriptionEventProcessorTest {
                 micrometerMetricsCounters);
 
         inOrder.verify(micrometerMetricsCounters).incrementEventsProcessedCount(source, component);
-        inOrder.verify(transactionHandler).rollback(userTransaction);
+        inOrder.verify(transactionHandler).rollback();
 
         verify(micrometerMetricsCounters, never()).incrementEventsSucceededCount(source, component);
         verify(micrometerMetricsCounters, never()).incrementEventsIgnoredCount(source, component);
@@ -680,7 +676,7 @@ public class SubscriptionEventProcessorTest {
                 micrometerMetricsCounters);
 
         inOrder.verify(micrometerMetricsCounters).incrementEventsProcessedCount(source, component);
-        inOrder.verify(transactionHandler).rollback(userTransaction);
+        inOrder.verify(transactionHandler).rollback();
         inOrder.verify(streamErrorStatusHandler).onStreamProcessingFailure(eventJsonEnvelope, nullPointerException, source, component, streamUpdateContext);
 
         verify(micrometerMetricsCounters, never()).incrementEventsSucceededCount(source, component);